+++ Razor/Scripts/Engine/Interpreter.cs
@@ -1276,6 +1276,24 @@ namespace Assistant.Scripts.Engine
             scope.ClearIgnore();
         }
         
+        public static void AddIgnoreRange(List<Serial> serials, bool global = true)
+        {
+            Scope scope = global ? _scope : _currentScope;
+            scope.AddIgnoreRange(serials);
+        }
+
+        public static void RemoveIgnore(Serial serial, bool global = true)
+        {
+            Scope scope = global ? _scope : _currentScope;
+            scope.RemoveIgnore(serial);
+        }
+
+        public static void RemoveIgnoreRange(List<Serial> serials, bool global = true)
+        {
+            Scope scope = global ? _scope : _currentScope;
+            scope.RemoveIgnoreRange(serials);
+        }
+        
         public static bool CheckIgnored(Serial serial, bool global = true)
         {
             Scope scope = global ? _scope : _currentScope;